harden action gate exit-code capture, rescope live ingest on ignore-rule edits, usage errors exit 2
- action.yml: capture the deslop exit code with an errexit-proof guard so
  the report/upload/gate steps actually run on a duplication breach;
  contract test and self-test breach assertions added
- live loop: forward .gitignore/.ignore edits through the watcher, treat
  them as corpus-reshaping in the session, and evict newly-ignored files
  (issue 287 follow-up), covered by session-level and full-loop tests
- split binding_drift.rs into mod.rs + scan.rs (500-line rule)
- consolidate the distinct-visible-paths screen into one shared helper
- CLI: post-parse usage errors now exit 2 like clap rejections
- drop stray pyproject.toml; document the exclusion-funded part of the
  12.2 duplication ratchet
