+++ action.yml
@@ -1,6 +1,6 @@
 # Deslop — GitHub Marketplace action. [ACTION-METADATA]
 #
-# Installs the published deslop CLI for the current runner, analyses the
+# Installs the published deslop CLI for the current runner, analyzes the
 # workspace, renders the reports, and fails the job when repo-wide duplication
 # breaches the threshold. Exit codes are owned by docs/specs/pipeline.md
 # [EXIT-CODES]; this action surfaces them, it does not reinterpret them.
@@ -14,7 +14,7 @@ branding:
 
 inputs:
   path:
-    description: Directory to analyse.
+    description: Directory to analyze.
     required: false
     default: "."
   version:
@@ -27,7 +27,7 @@ inputs:
     default: ""
   fail-over:
     description: >-
-      Duplication percentage above which the job fails. Leave unset to honour
+      Duplication percentage above which the job fails. Leave unset to honor
       `[threshold] max_duplication_percent` in .deslop.toml. Must be a finite
       number in [0, 100].
     required: false
@@ -77,7 +77,7 @@ inputs:
 
 outputs:
   duplication-percent:
-    description: Repo-wide duplicated lines as a percentage of analysed lines.
+    description: Repo-wide duplicated lines as a percentage of analyzed lines.
     value: ${{ steps.report.outputs.duplication-percent }}
   cluster-count:
     description: Number of clusters contributing to the duplicated line count.
@@ -145,9 +145,12 @@ runs:
         ARCHIVE: ${{ steps.resolve.outputs.archive }}
         STAGE: ${{ steps.resolve.outputs.stage }}
 
-    # `set -e` is deliberately absent: a non-zero deslop status is the product,
-    # not a failure of this step. The status is captured and re-raised by the
-    # gate step below so the reports are read and uploaded first.
+    # A non-zero deslop status is the product, not a failure of this step.
+    # GitHub runs composite `shell: bash` steps under an injected `-e`
+    # (`bash -eo pipefail`), and a later `set` cannot clear it — so the status
+    # is captured with an `||` guard, which errexit exempts. The step then
+    # always succeeds, the report step publishes the measurements, the
+    # artifact uploads, and the gate step below re-raises the status.
     - name: Run deslop
       id: run
       shell: bash
@@ -160,8 +163,9 @@ runs:
         if [ "${NOJSON}" = "true" ]; then args+=(--nojson); fi
         if [ "${NOTEXT}" = "true" ]; then args+=(--notext); fi
         if [ "${NOHTML}" = "true" ]; then args+=(--nohtml); fi
-        deslop "${args[@]}"
-        echo "exit-code=$?" >> "${GITHUB_OUTPUT}"
+        status=0
+        deslop "${args[@]}" || status=$?
+        echo "exit-code=${status}" >> "${GITHUB_OUTPUT}"
       env:
         SCAN_PATH: ${{ inputs.path }}
         MIN_NODES: ${{ inputs.min-nodes }}
@@ -201,7 +205,7 @@ runs:
       shell: bash
       run: |
         case "${EXIT_CODE}" in
-          3) echo "::error::Deslop gate failed — ${MEASURED}% of analysed lines are duplicated, ceiling is ${CEILING}%. The full report is in the ${ARTIFACT_NAME} artifact." ;;
+          3) echo "::error::Deslop gate failed — ${MEASURED}% of analyzed lines are duplicated, ceiling is ${CEILING}%. The full report is in the ${ARTIFACT_NAME} artifact." ;;
           1) echo "::error::Deslop could not complete the analysis (exit 1). Check the scan path, the .deslop.toml config, and the log output above." ;;
           2) echo "::error::Deslop rejected its arguments (exit 2). Check the action inputs — fail-over must be a finite number in [0, 100] and cannot be combined with no-fail-over." ;;
           *) echo "::error::Deslop exited with unexpected status ${EXIT_CODE}." ;;
