+++ crates/deslop-core/src/pipeline/session/change.rs
@@ -106,17 +106,20 @@ impl PipelineSession {
         Ok(())
     }
 
-    /// Drops every live file the reloaded exclusion config now
-    /// matches. Returns the number of dropped paths.
+    /// Drops every live file the reloaded exclusion config — or the
+    /// rebuilt ignore matcher — now matches, so a `.deslop.toml` or
+    /// `.gitignore` edit evicts exactly what a fresh discovery would
+    /// prune (ignore-rule parity #287). Returns the number of dropped
+    /// paths.
     fn drop_newly_excluded(&mut self) -> usize {
         let doomed: Vec<PathBuf> = self
             .live_paths
             .iter()
             .filter_map(|(file_id, path)| {
                 let language = self.file_languages.get(file_id).copied();
-                self.exclusion
-                    .is_excluded(path, language)
-                    .then(|| path.clone())
+                let excluded = self.exclusion.is_excluded(path, language)
+                    || self.ignore_matcher.is_ignored(path);
+                excluded.then(|| path.clone())
             })
             .collect();
         for path in &doomed {
